Found solution to horrible linux bug that turned out to be feature
I encountered this problem where i could only receive packets on a
single internet connection at a time: whichever had been configured as
the default route.  the packets showed up in tshark and hit the
PREROUTING stages of iptables, but no matter what i did i couldn't get
them to reach a socket.  i tried raw sockets, SO_BINDTODEVICE, you
name it... but the packets kept mysteriously vanishing.

After many hours of frustration I finally realized that the kernel's
IP spoofing profection (reverse path filtering) was the problem so I
disabled it.  Please note that most tutorials that show how to do this
on the web are wrong because it needs to be disabled manually for each
network device.